Upon launching main program, it will get current settings from video feed (v4lctl) and will save them to restore_v4lctl_settings.sh. Will also set some hardcoded settings for v4lctl (these attributes are what makes people say "vision looks different on different pcs". After testing this with 5pcs (2 at pitch0, 3 at pitch1), I am quite confident in saying, that now vision looks the same for the same pitch on different computers (any minor differences should be accounted to monitor settings). Thus this also reduces our calibrations.json file to "default" case. I've calibrated vision for both pitches now [plate, dot and red].
Bash file will be created only if any of the settings was different than to the ones, that are being set.
Ideally, it should be run only after we are finished with that computer (being all mindfull about others and stuff, or we could not run it and maybe mess with other people's calibration? :D).
To run it:
cd to main dir (where .sh file is)
$ chmod 755 restore_v4lctl_settings.sh
$ ./restore_v4lctl_settings.sh